UserService/Handlers/GetAllUsersQueryHandler.cs
using MediatR;
using UserService.Models;
using UserService.Queries;
using UserService.Data;

namespace UserService.Handlers
{
    public class GetAllUsersQueryHandler : IRequestHandler<GetAllUsersQuery, IEnumerable<User>>
    {
        private readonly IUnitOfWork _unitOfWork;

        public GetAllUsersQueryHandler(IUnitOfWork unitOfWork)
        {
            _unitOfWork = unitOfWork;
        }

        public async Task<IEnumerable<User>> Handle(GetAllUsersQuery request, CancellationToken cancellationToken)
        {
            return await _unitOfWork.Users.GetAllUsersAsync();
        }
    }
}
